Add support for CelebA test/train split
The current training routine is training on all the CelebA
images, including the canonical CelebA validation and
test data. This change adds support for CelebA splits
and by default only trains on only the CelebA training
images.

This was done by moving the downloaded data into an
"images" subdirectory, and then creating subdirectories
of splits with relative symlinks to the original files.
This format required only light changes to the config
and data_loader.

Note that honoring the standard splits in generative
modeling is important for a few reasons. Perhaps most
importantly, it allows for fair comparisons across papers
when visually inspecting image samples; it is unfair
to compare models trained on 200k images versus those
only exposed to a 150k subset. But also relevant for
models with an encoder like BEGAN - it allows one to
evaluate a trained model by performing reconstructions
on out of sample validation data.