Added autoencoded image output when run in test mode
Running in test mode was not producing any output, so I added
an initial implementation of Trainer.test(). This simply
autoencodes a batch of inputs, and saves the inputs
as well as the autoencoded outputs.

This can be combined with '--split valid' to test
autoencoding on validation data. To make these tests
more repeatable, shuffling was disabled in test
mode and the images are sort ordered.